월드컵 수정페이지 수정
월드컵 사진수정,삭제기능 추가(수정,삭제 시 기존 이미지 파일 삭제) 월드컵 영상 링크 수정,삭제 기능 추가, 영상 링크 저장 시 유튜브 영상 코드를 통한 썸네일 이미지 표시 추가, 월드컵 제목,설명,공개 기능 수정 추가

+++ src/main/java/com/world/cup/repository/WorldcupRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface WorldcupRepository extends JpaRepository<Worldcup, Long>, WorldcupListRepository {
+public interface WorldcupRepository extends JpaRepository<Worldcup, Integer>, WorldcupListRepository {
     @Query("select w from Worldcup w where w.worldcupNum = :worldcupNum")
     Object getWorldcupByWorldcupNum(@Param("worldcupNum") int worldcupNum);
